Modifying DB column datatype to float
- Modifying DB column datatype from double to float in tables
  learner_report_activity and learner_report

+++ db/migrate/20120712065913_add_complete_percent_to_report_learner.rb
@@ -6,7 +6,7 @@ class AddCompletePercentToReportLearner < ActiveRecord::Migration
   end
 
   def self.up
-    add_column :report_learners, :complete_percent, :double
+    add_column :report_learners, :complete_percent, :float
     Report::Learner.reset_column_information
     learners = Report::Learner.all
     learners.each do |learner|
